requirements.txt
# requirements.txt
requests
flask
ics
beautifulsoup4
lxml
pyjwt